Module bio::alignment::distance

Various subroutines for computing a distance between sequences. Features both scalar and efficient vectorized distance functions with SIMD.

Modules

simd
String distance routines accelerated with Single Instruction Multiple Data (SIMD) intrinsics.

Functions

hamming
Compute the Hamming distance between two strings. Complexity: O(n).

levenshtein
Compute the Levenshtein (or Edit) distance between two strings. Complexity: O(n * m) with n and m being the length of the given texts.