Struct bio::alignment::poa::Poa

pub struct Poa<F: MatchFunc> {
    pub graph: POAGraph,
    // some fields omitted
}

A partially ordered alignment graph

A directed acyclic graph datastructure that represents the topology of a traceback matrix.

Fields
graph: POAGraph
Implementations

impl<F: MatchFunc> Poa<F>

pub fn new(scoring: Scoring<F>, graph: POAGraph) -> Self

Create a new aligner instance from the directed acyclic graph of another.

Arguments

scoring - the score struct

poa - the partially ordered reference alignment

pub fn from_string(scoring: Scoring<F>, seq: TextSlice<'_>) -> Self

Create a new POA graph from an initial reference sequence and alignment penalties.

Arguments

scoring - the score struct

reference - a reference TextSlice to populate the initial reference graph

pub fn global(&self, query: TextSlice<'_>) -> Traceback

A global Needleman-Wunsch aligner on partially ordered graphs.

Arguments

query - the query TextSlice to align against the internal graph member

pub fn edges(&self, aln: Alignment) -> Vec<usize>

Experimental: return sequence of traversed edges

Only supports alignments for sequences that have already been added, so all operations must be Match.

pub fn add_alignment(&mut self, aln: &Alignment, seq: TextSlice<'_>)

Incorporate a new sequence into a graph from an alignment

Arguments

aln - The alignment of the new sequence to the graph

seq - The sequence being incorporated