Struct bio::alphabets::RankTransform

pub struct RankTransform {
    pub ranks: SymbolRanks,
}

Tools based on transforming the alphabet symbols to their lexicographical ranks.

Lexicographical rank is computed using u8 representations, i.e. ASCII codes, of the input characters. For example, assuming that the alphabet consists of the symbols A, C, G, and T, this will yield ranks 0, 1, 2, 3 for them, respectively.

RankTransform can be used in to perform bit encoding for texts over a given alphabet via bio::data_structures::bitenc.

Fields
ranks: SymbolRanks
Implementations

impl RankTransform

pub fn new(alphabet: &Alphabet) -> Self

Construct a new RankTransform.

Complexity: O(n), where n is the number of symbols in the alphabet.

Example

use bio::alphabets;

let dna_alphabet = alphabets::Alphabet::new(b"acgtACGT");
let dna_ranks = alphabets::RankTransform::new(&dna_alphabet);

pub fn get(&self, a: u8) -> u8

Get the rank of symbol a.

This method panics for characters not contained in the alphabet.

Complexity: O(1)

Example

use bio::alphabets;

let dna_alphabet = alphabets::Alphabet::new(b"acgtACGT");
let dna_ranks = alphabets::RankTransform::new(&dna_alphabet);
assert_eq!(dna_ranks.get(65), 0); // "A"
assert_eq!(dna_ranks.get(116), 7); // "t"

pub fn transform<C, T>(&self, text: T) -> Vec<u8> where
C: Borrow<u8>,
T: IntoIterator<Item = C>,

Transform a given text into a vector of rank values.

Complexity: O(n), where n is the length of the text.

Example

use bio::alphabets;

let dna_alphabet = alphabets::Alphabet::new(b"ACGTacgt");
let dna_ranks = alphabets::RankTransform::new(&dna_alphabet);
let text = b"aAcCgGtT";
assert_eq!(dna_ranks.transform(text), vec![4, 0, 5, 1, 6, 2, 7, 3]);

pub fn qgrams<C, T>(&self, q: u32, text: T) -> QGrams<'_, C, T::IntoIter> where
C: Borrow<u8>,
T: IntoIterator<Item = C>,

Iterate over q-grams (substrings of length q) of given text. The q-grams are encoded as usize by storing the symbol ranks in log2(|A|) bits (with |A| being the alphabet size).

If q is larger than usize::BITS / log2(|A|), this method fails with an assertion.

Complexity: O(n), where n is the length of the text.

Example

use bio::alphabets;

let dna_alphabet = alphabets::Alphabet::new(b"ACGTacgt");
let dna_ranks = alphabets::RankTransform::new(&dna_alphabet);

let q_grams: Vec<usize> = dna_ranks.qgrams(2, b"ACGT").collect();
assert_eq!(q_grams, vec![1, 10, 19]);

pub fn alphabet(&self) -> Alphabet

Restore alphabet from transform.

Complexity: O(n), where n is the number of symbols in the alphabet.

Example

use bio::alphabets;

let dna_alphabet = alphabets::Alphabet::new(b"acgtACGT");
let dna_ranks = alphabets::RankTransform::new(&dna_alphabet);
assert_eq!(dna_ranks.alphabet().symbols, dna_alphabet.symbols);

pub fn get_width(&self) -> usize

Compute the number of bits required to encode the largest rank value.

For example, the alphabet b"ACGT" with 4 symbols has the maximal rank 3, which can be encoded in 2 bits.

This value can be used to create a data_structures::bitenc::BitEnc bit encoding tailored to the given alphabet.

Complexity: O(n), where n is the number of symbols in the alphabet.

Example

use bio::alphabets;

let dna_alphabet = alphabets::Alphabet::new(b"ACGT");
let dna_ranks = alphabets::RankTransform::new(&dna_alphabet);
assert_eq!(dna_ranks.get_width(), 2);
let dna_n_alphabet = alphabets::Alphabet::new(b"ACGTN");
let dna_n_ranks = alphabets::RankTransform::new(&dna_n_alphabet);
assert_eq!(dna_n_ranks.get_width(), 3);