Struct bio::data_structures::fmindex::FMDIndex

pub struct FMDIndex<DBWT: Borrow<BWT>, DLess: Borrow<Less>, DOcc: Borrow<Occ>> { /* fields omitted */ }

The FMD-Index for linear time search of supermaximal exact matches on forward and reverse strand of DNA texts (Li, 2012).

Methods

impl<DBWT: Borrow<BWT>, DLess: Borrow<Less>, DOcc: Borrow<Occ>> FMDIndex<DBWT, DLess, DOcc>

pub fn smems(&self, pattern: &[u8], i: usize) -> Vec<BiInterval>

Find supermaximal exact matches of given pattern that overlap position i in the pattern. Complexity O(m) with pattern of length m.

Example

use bio::alphabets::dna;
use bio::data_structures::fmindex::{FMIndex, FMDIndex};
use bio::data_structures::suffix_array::suffix_array;
use bio::data_structures::bwt::{bwt, less, Occ};

let text = b"ATTC$GAAT$";
let alphabet = dna::n_alphabet();
let sa = suffix_array(text);
let bwt = bwt(text, &sa);
let less = less(&bwt, &alphabet);
let occ = Occ::new(&bwt, 3, &alphabet);
let fm = FMIndex::new(&bwt, &less, &occ);
let fmdindex = FMDIndex::from(fm);

let pattern = b"ATT";
let intervals = fmdindex.smems(pattern, 2);

let forward_positions = intervals[0].forward().occ(&sa);

let revcomp_positions = intervals[0].revcomp().occ(&sa);

assert_eq!(forward_positions, [0]);
assert_eq!(revcomp_positions, [6]);

pub fn init_interval_with(&self, a: u8) -> BiInterval

Initialize interval with given start character.

pub fn init_interval(&self) -> BiInterval

Initialize interval for empty pattern. The interval points at the whole suffix array.

pub fn backward_ext(&self, interval: &BiInterval, a: u8) -> BiInterval

Backward extension of given interval with given character.

pub fn forward_ext(&self, interval: &BiInterval, a: u8) -> BiInterval