Struct bio::data_structures::fmindex::FMIndex

pub struct FMIndex<'a> {
    // some fields omitted
}

The Fast Index in Minute space (FM-Index, Ferragina and Manzini, 2000) for finding suffix array intervals matching a given pattern.

Methods

impl<'a> FMIndex<'a>

fn new(bwt: &'a BWT, k: usize, alphabet: &Alphabet) -> Self

Construct a new instance of the FM index.

Arguments

bwt - the BWT

k - the sampling rate of the occ array: every k-th entry will be stored (higher k means less memory usage, but worse performance)

alphabet - the alphabet of the underlying text, omitting the sentinel

fn backward_search<'b, P: Iterator<Item=&'b u8> + DoubleEndedIterator>(&self, pattern: P) -> Interval

Perform backward search, yielding suffix array interval denoting exact occurences of the given pattern of length m in the text. Complexity: O(m).

Arguments

pattern - the pattern to search

Example

use bio::data_structures::bwt::bwt;
use bio::data_structures::fmindex::FMIndex;
use bio::data_structures::suffix_array::suffix_array;
use bio::alphabets::dna;

let text = b"GCCTTAACATTATTACGCCTA$";
let alphabet = dna::alphabet();
let pos = suffix_array(text);
let bwt = bwt(text, &pos);
let fm = FMIndex::new(&bwt, 3, &alphabet);

let pattern = b"TTA";
let sai = fm.backward_search(pattern.iter());

let occ = sai.occ(&pos);

assert_eq!(occ, [3, 12, 9]);