Struct bio::data_structures::interval_tree::Entry

pub struct Entry<'a, N: 'a, D: 'a> {
    // some fields omitted
}

Methods

impl<'a, N: 'a, D: 'a> Entry<'a, N, D>

fn data(&self) -> &'a D

Get a reference to the data for this entry

fn interval(&self) -> &'a Range<N>

Get a reference to the interval for this entry