Struct bio::data_structures::rank_select::RankSelect

pub struct RankSelect {
    // some fields omitted
}

Methods

impl RankSelect

fn new(bits: Bitv, k: usize) -> RankSelect

Create a new instance.

Arguments

bits - A bit vector.

k - Determines the size (k * 32 bits) of the superblocks. A small k means faster rank query times at the expense of using more space and slower select query times.

fn rank(&self, i: usize) -> Option<u32>

Get the rank of a given bit. Complexity: O(k).

Arguments

i - Position of the bit to determine the rank for.

fn select(&self, j: u32) -> Option<usize>

Get the smallest bit with a given rank. Complexity: O(log (n / k) + k).

Arguments

j - The rank to find the smallest bit for.