Module bio::io

Readers and writers for common bioinformatics file formats.

Modules

bed

BED format reading and writing.

fasta

Structs and trait to read and write files in FASTA format.

fastq

Structs and trait to read and write files in FASTQ format.

gff

GFF3 format reading and writing. GFF2 is currently not supported.