Struct bio::io::bed::Record

pub struct Record { /* fields omitted */ }

A BED record as defined by BEDtools (http://bedtools.readthedocs.org/en/latest/content/general-usage.html)

Methods

impl Record

pub fn new() -> Self

Create a new BED record.

pub fn chrom(&self) -> &str

Chromosome of the feature.

pub fn start(&self) -> u64

Start position of feature (0-based).

pub fn end(&self) -> u64

End position of feature (0-based, not included).

pub fn name(&self) -> Option<&str>

Name of the feature.

pub fn score(&self) -> Option<&str>

Score of the feature.

pub fn strand(&self) -> Option<Strand>

Strand of the feature.

pub fn aux(&self, i: usize) -> Option<&str>

Access auxilliary fields after the strand field by index (counting first field (chromosome) as 0).

pub fn set_chrom(&mut self, chrom: &str)

Set chromosome.

pub fn set_start(&mut self, start: u64)

Set start of feature.

pub fn set_end(&mut self, end: u64)

Set end of feature.

pub fn set_name(&mut self, name: &str)

Set name.

pub fn set_score(&mut self, score: &str)

Set score.

pub fn push_aux(&mut self, field: &str)

Add auxilliary field. This has to happen after name and score have been set.