Struct bio::io::fastq::Reader

pub struct Reader<R: Read> {
    // some fields omitted
}

A FastQ reader.

Methods

impl Reader<File>

fn from_file<P: AsRef<Path>>(path: P) -> Result<Self>

Read from a given file.

impl<R: Read> Reader<R>

fn new(reader: R) -> Self

Read from a given io::Read.

fn read(&mut self, record: &mut Record) -> Result<()>

Read into a given record. Returns an error if the record in incomplete or syntax is violated. The content of the record can be checked via the record object.

fn records(self) -> Records<R>

Return an iterator over the records of this FastQ file.