bio::io

Module fastx

Traits and utilities to read and write files in the FASTA and FASTQ format interchangably

Files in the FASTA and FASTQ format have several common fields: ID, Sequence, and Description. These utilities can be used to implement algorithms that only require these fields for files that are in either format.

This module serves two use cases:

Implementing functions that can be used generically with FASTA/FASTQ records. In this use case the type may be known at compile time by the caller of your function.

Processing data that may be either in the FASTA/FASTQ format. In this use case the type cannot be known at compile time and you may or may not want to treat FASTA/FASTQ data differently.

Generic Implementation Examples

Common Statistics

In this example, we implement a count_bases function that for supports both the FASTA and FASTQ format.

use bio::io::{fasta, fastq, fastx};
use std::io;

fn count_bases<T, E, I>(records: I) -> Result<usize, E>
where
    T: fastx::Record,
    E: std::error::Error,
    I: fastx::Records<T, E>,
{
    let mut nb_bases = 0;
    for result in records {
        let record = result?;
        nb_bases += record.seq().len();
    }
    Ok(nb_bases)
}

let mut raw_reader = io::Cursor::new(
    b">id desc
ACTG
",
);

match fastx::get_kind_seek(&mut raw_reader) {
    Ok(fastx::Kind::FASTA) => {
        let mut reader = fasta::Reader::new(raw_reader);
        let nb_bases = count_bases(reader.records()).unwrap();
        println!("Number of bases: {}", nb_bases);
    }
    Ok(fastx::Kind::FASTQ) => {
        let mut reader = fastq::Reader::new(raw_reader);
        let nb_bases = count_bases(reader.records()).unwrap();
        println!("Number of bases: {}", nb_bases);
    }
    _ => println!("Encountered an error"),
}

Filtration

In this example, we define an at_least_n_bases function that can filter FASTA or FASTQ records based on their sequence lengths. It works seemlessly with fasta::Records as if it was implemented just for them. In a realistic scenario this function might be defined in a library so callers could use it with both FASTA and FASTQ files as needed.

use bio::io::{fasta, fastq, fastx};
use std::io;
use std::io::BufReader;

fn at_least_n_bases<T, E, I>(records: I, n: usize) -> impl Iterator<Item = Result<T, E>>
where
    T: fastx::Record,
    E: std::error::Error,
    I: fastx::Records<T, E>,
{
    records.filter(move |rr| match rr {
        Ok(r) => r.seq().len() > n,
        _ => true,
    })
}

let mut reader = fasta::Reader::new(BufReader::new(io::stdin()));
let mut writer = fasta::Writer::new(io::stdout());

for record in at_least_n_bases(reader.records(), 10) {
    writer.write_record(&record.unwrap());
}

Dynamic Type Examples

If the type of a record is not known at compile time you can represent it with fastx::EitherRecord. This type is an enum containing either a fasta::Record or a fastq::Record. There are also utility functions defined on the enum so you can work with them without converting to the underlying type.

Note that using dynamic types like this incurs a slight performance cost around 5-10% when compared to the typing based approach or parsing a single type directly.

Parsing data of either the FASTA of FASTQ type

use bio::io::fastx::{EitherRecords, Record};
use std::io;
use std::io::BufReader;
use std::str;

let mut records = EitherRecords::from(BufReader::new(io::stdin()));
while let Some(Ok(record)) = records.next() {
    println!("id  : {}", record.id());
    println!("desc: {}", record.desc().unwrap_or("none"));
    println!("seq : {}", str::from_utf8(record.seq()).unwrap_or("error"));
    // Add a default quality in case we have a FASTA record
    let default_qual = vec![b'I'; record.seq().len()];
    let qual = record.qual().unwrap_or(&default_qual);
    println!("qual: {}", str::from_utf8(qual).unwrap_or("error"));
    println!("")
}

Utility Examples

Type Detection

This module provides utilities for detecting if some data is a FASTA/FASTQ.

use bio::io::fastx::{get_kind, get_kind_file, get_kind_seek, Kind};
use std::fs;
use std::fs::File;
use std::io;
use std::io::prelude::*;

// From a Read

fn from_read() -> io::Result<Kind> {
    let reader = io::stdin();
    let (mut new_reader, kind) = get_kind(reader)?;
    println!("{}", kind);
    // Read from start of your old reader
    let mut buf = [0u8; 8];
    new_reader.read(&mut buf)?;
    Ok(kind)
}

// From a Read + Seek

fn from_read_seek() -> io::Result<Kind> {
    let mut read_seek = io::Cursor::new(
        b">id desc
ACTG
",
    );

    get_kind_seek(&mut read_seek)
}

// From a file path

fn from_file_path() -> io::Result<Kind> {
    get_kind_file("foo.fasta")
}

Structs

EitherRecords

Enums

EitherRecord

Error

Kind

Traits

Record

Records

Functions

get_kind

Determine whether a Read is a FASTA or FASTQ.

get_kind_detailed

Determine whether a Read is a FASTA or FASTQ.

get_kind_file

Determine whether a file is a FASTA or FASTQ.

get_kind_seek

Determine whether a Read + Seek is a FASTA or FASTQ.