bio::io::fastx

Function get_kind

pub fn get_kind<R: Read>(reader: R) -> Result<(Chain<Cursor<[u8; 1]>, R>, Kind)>

Determine whether a Read is a FASTA or FASTQ.

This function is a wrapper around get_kind_detailed for more convenient use if you don’t want to:

Access your input Read in the event of an error

Explicitly handle errors resulting from reading from your Read differently from errors resulting from an invalid FASTA/FASTQ

You should also only use this function if your Read does not implement Seek othwerwise get_kind_seek is more convenient.

This method takes ownership of the Read and returns a new Read that is equivalent in content to the input Read. This allows you to pass the returned Read#Read) directly into fasta::Reader::new or fastq::Reader::new.

Example

use bio::io::fastx::{get_kind, Kind};
use bio::io::{fasta, fastq};
use std::io;

fn count_records() -> io::Result<usize> {
    let (reader, kind) = get_kind(io::stdin())?;
    match kind {
        Kind::FASTA => Ok(fasta::Reader::new(reader).records().count()),
        Kind::FASTQ => Ok(fastq::Reader::new(reader).records().count()),
    }
}