bio::io::fastx

Function get_kind_detailed

pub fn get_kind_detailed<R: Read>(
    reader: R,
) -> Result<(Chain<Cursor<[u8; 1]>, R>, Result<Kind>), (R, Error)>

Determine whether a Read is a FASTA or FASTQ.

You should only use this function if you would like to:

Access your input Read in the event of an error

Explicitly handle errors resulting from reading from your Read differently from errors resulting from an invalid FASTA/FASTQ

Otherwise get_kind will be more convenient.

You should also only use this function if your Read does not implement Seek othwerwise get_kind_seek is more convenient.

This method takes ownership of the input Read. It reads from the Read to determine whether the data is in the FASTA or FASTQ format. If this read fails the function returns an Err containing a tuple of the input read and the error it encountered (return Err((read, err))). Note that in this case there are no guarantee about the position in the returned reader and some data may become unreadable if the position was advanced. Based on the contents of the error the caller can determine how best to recover. If the read is successful this function will always return Ok for the outer Result with a tuple containing a new Read and a Result containing the fastx::Kind if the data was a valid FASTA or FASTQ file or an io::Error if the format was invalid (ex. Ok((new_reader), Ok(Kind::FASTA))). The new Read has the character this function read prepended to it so the resulting Read should be equivalent in content to the input Read.

Example

use bio::io::fastx::{get_kind_detailed, Kind};
use bio::io::{fasta, fastq};
use std::io;

fn print_type() {
    let res = get_kind_detailed(io::stdin());
    match res {
        Ok((_reader, Ok(Kind::FASTA))) => println!("{}", Kind::FASTA),
        Ok((_reader, Ok(Kind::FASTQ))) => println!("{}", Kind::FASTQ),
        Ok((mut reader, Err(e))) => {
            println!("Error determining FASTA/FASTQ: {}", e);
            println!("Data:");
            io::copy(&mut reader, &mut io::stdout());
        }
        Err((mut reader, e)) => {
            println!("Encountered an error while determining FASTA/FASTQ: {}", e);
            println!("Remaing data:");
            io::copy(&mut reader, &mut io::stdout());
        }
    }
}