bio::io::fastx

Function get_kind_seek

pub fn get_kind_seek<R: Read + Seek>(reader: &mut R) -> Result<Kind>

Determine whether a Read + Seek is a FASTA or FASTQ.

The benefit of this this function compared to get_kind is that this function does not take ownership of the Read so it can be slightly more convenient to use.