Enum bio::io::gff::GffType

pub enum GffType {
    GFF3,
    GFF2,
    GTF2,
    Any(u8, u8),
}

GffType

We have three format in the GFF family. The change is in the last field of GFF. For each type we have key value separator and field separator

Variants
GFF3

Attribute format is key1=value, key2=value

GFF2

Attribute format is key1 value; key2 value

GTF2

Same as GFF2 just possible keyword and possible value change

Any(u8, u8)

Any, first field of tuple is key value separator, second is field separator