Struct bio::pattern_matching::bom::BOM

pub struct BOM {
    // some fields omitted
}

Backward oracle matching algorithm.

Methods

impl BOM

fn new(pattern: &[u8]) -> Self

Create a new instance for a given pattern.

fn find_all<'a>(&'a self, text: &'a [u8]) -> Matches

Find all matches of the pattern in the given text. Matches are returned as an iterator over start positions.