bio::pattern_matching::myers

Trait BitVec

pub trait BitVec:
    Copy
    + Debug
    + Binary
    + Default
    + Add
    + Sub
    + BitOr
    + BitOrAssign
    + BitAnd
    + BitXor
    + Not
    + Shl<usize>
    + ShlAssign<usize>
    + ShrAssign<usize>
    + PrimInt
    + WrappingAdd
    + ToPrimitive
    + FromPrimitive
    + AsPrimitive<usize>
    + AsPrimitive<i64> {
    type DistType: DistType;
}

This trait must be implemented for integer types serving as bit vectors. Only unsigned integers will work correctly.

Required Associated Types

type DistType: DistType

Type that should be used to store the distance score when using the simple Myers algorithm (not the block-based one, which always uses usize).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.