bio::pattern_matching::myers

Trait DistType

pub trait DistType:
    Copy
    + Debug
    + Default
    + AddAssign
    + SubAssign
    + PrimInt
    + FromPrimitive
    + ToPrimitive
    + AsPrimitive<usize>
    + AsPrimitive<i64>
    + WrappingAdd
    + Sub<Output = Self> { }

Trait for types used to store the edit distance when using the simple Myers algorithm (not the block-based one, which always uses usize).

For all currently implemented BitVec types, the maximum possible distance can be stored in u8. Custom implementations using bigger integers can adjust DistType to hold bigger numbers.

Note: For the traceback algorithm to work, the maximum possible edit distance (which is the number of bits of the BitVec type) should be smaller than <DistType>::MAX. A hypothetical BitVec type of u256 should have DistType = u16, since u8 cannot represent numbers > 255.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.