Struct bio::pattern_matching::shift_and::ShiftAnd

pub struct ShiftAnd {
    // some fields omitted
}

ShiftAnd algorithm.

Methods

impl ShiftAnd

fn new<'a, P: IntoTextIterator<'a>>(pattern: P) -> Self where P::IntoIter: ExactSizeIterator

Create new ShiftAnd instance from a given pattern.

fn find_all<'a, I: IntoTextIterator<'a>>(&'a self, text: I) -> Matches<I::IntoIter>

Find all matches of pattern in the given text. Matches are returned as an iterator over start positions.