Struct bio::pattern_matching::ukkonen::Ukkonen

pub struct Ukkonen<F> where F: Fn(u8, u8) -> u32 {
    // some fields omitted
}

Ukkonens algorithm.

Methods

impl<F> Ukkonen<F> where F: Fn(u8, u8) -> u32

fn with_capacity(m: usize, cost: F) -> Self

Initialize algorithm with given capacity and cost function.

fn find_all_end<'a, T: IntoTextIterator<'a>>(&'a mut self, pattern: TextSlice<'a>, text: T, k: usize) -> Matches<F, T::IntoIter>

Find all matches between pattern and text with up to k errors. Matches are returned as an iterator over pairs of end position and distance.