Struct bio::seq_analysis::orf::Finder

pub struct Finder {
    // some fields omitted
}

An implementation of a naive algorithm finder

Methods

impl Finder

fn new<'a>(start_codons: Vec<&'a [u8; 3]>, stop_codons: Vec<&'a [u8; 3]>, min_len: usize) -> Self

Create a new instance of a finder for the given start and stop codons and a particular length

fn find_all<'a, I: IntoTextIterator<'a>>(&'a self, seq: I) -> Matches<I::IntoIter>

Find all orfs in the given sequence