Module bio::stats

Mathematical and statistical tools.

Re-exports

pub use crate::stats::probs::LogProb;

pub use crate::stats::probs::PHREDProb;

pub use crate::stats::probs::Prob;

Modules

bayesian
Utilities for Bayesian statistics.

combinatorics
Combinations with and without replacement.

hmm
An implementation of Hidden Markov Models in Rust.

pairhmm
This module contains the implementation of a classic PairHMM as described in Durbin, R., Eddy, S., Krogh, A., & Mitchison, G. (1998). Biological Sequence Analysis. Current Topics in Genome Analysis 2008. http://doi.org/10.1017/CBO9780511790492. It also contains a modified variant HomopolyPairHMM with additional homopolymer states suited for dealing with homopolymer runs in sequencing as often encountered in Oxford Nanopore sequencing data.

probs
Handling log-probabilities. Log probabilities are an important tool to deal with probabilities in a numerically stable way, in particular when having probabilities close to zero.