Module bio::stats::bayesian

Utilities for Bayesian statistics.

Re-exports

pub use self::bayes_factors::BayesFactor;

pub use self::model::Model;

Modules

bayes_factors

model
A trait system for Bayesian statistical modelling.

Functions

expected_fdr
For each of the hypothesis tests given as posterior error probabilities (PEPs, i.e. the posterior probability of the null hypothesis), estimate the FDR for the case that all null hypotheses with at most this PEP are rejected. FDR is calculated as presented by Müller, Parmigiani, and Rice, "FDR and Bayesian Multiple Comparisons Rules" (July 2006). Johns Hopkin's University, Dept. of Biostatistics Working Papers. Working Paper 115.