Function bio::stats::bayesian::expected_fdr

pub fn expected_fdr(peps: &[LogProb]) -> Vec<LogProb>

For each of the hypothesis tests given as posterior error probabilities (PEPs, i.e. the posterior probability of the null hypothesis), estimate the FDR for the case that all null hypotheses with at most this PEP are rejected. FDR is calculated as presented by Müller, Parmigiani, and Rice, "FDR and Bayesian Multiple Comparisons Rules" (July 2006). Johns Hopkin's University, Dept. of Biostatistics Working Papers. Working Paper 115.

Returns

A vector of expected FDRs in the same order as the given PEPs.