Function bio::stats::hmm::backward

pub fn backward<O, M: Model<O>>(
    hmm: &M, 
    observations: &[O]
) -> (Array2<LogProb>, LogProb)

Execute the backward algorithm and return the backward probabilities as LogProb values and the resulting backward probability.

Arguments

hmm - the Model to run the backward algorithm on

observations - a slice of observation values to use in the algorithm

Result

The resulting pair (P, p) is the backward probability table (P[[s, o]] is the entry for state s and observation o) and the overall probability for observations (as LogProb).

Type Parameters

O - the observation type

M - type Model type