Struct bio::stats::hmm::univariate_continuous_emission::Model

pub struct Model<Dist: Continuous<f64, f64>> { /* private fields */ }

Implementation of a hmm::Model with emission values from univariate continuous distributions.

Log-scale probabilities are used for numeric stability.

Implementations

impl<Dist: Continuous<f64, f64>> Model<Dist>

pub fn new(
transition: Array2<LogProb>,
observation: Vec<Dist>,
initial: Array1<LogProb>
) -> Result<Self>

Construct new Hidden MarkovModel with the given transition, observation, and initial state matrices and vectors already in log-probability space.

pub fn with_prob(
transition: &Array2<Prob>,
observation: Vec<Dist>,
initial: &Array1<Prob>
) -> Result<Self>

Construct new Hidden MarkovModel with the given transition, observation, and initial state matrices and vectors already as Prob values.

pub fn with_float(
transition: &Array2<f64>,
observation: Vec<Dist>,
initial: &Array1<f64>
) -> Result<Self>

Construct new Hidden MarkovModel with the given transition, observation, and initial state matrices and vectors with probabilities as f64 values.