Module bio::stats::logprobs

Handling log-probabilities.

Reexports

pub use stats::{Prob, LogProb};

Functions

ln_1m_exp
Calculate log(1 - p) with p given in log space without loss of precision as described in http://cran.r-project.org/web/packages/Rmpfr/vignettes/log1mexp-note.pdf.

log_prob_add
Calculate the sum of the given probabilities in a numerically stable way (Durbin 1998).

log_prob_cumsum
Calcualte the cumulative sum of the given probabilities in a numerically stable way (Durbin 1998).

log_prob_sum
Calculate the sum of the given probabilities in a numerically stable way (Durbin 1998).

log_to_phred
Convert log scale probability to PHRED scale.

phred_to_log
Convert PHRED scale probability to log scale.