bio::stats::pairhmm

Trait StartEndGapParameters

pub trait StartEndGapParameters {
    // Required methods
    fn free_start_gap_x(&self) -> bool;
    fn free_end_gap_x(&self) -> bool;

    // Provided method
    fn prob_start_gap_x(&self, i: usize) -> LogProb { ... }
}

Trait for parametrization of PairHMM start and end gap behavior. This trait can be used to implement global and semiglobal alignments.

global: methods return false and LogProb::ln_zero().

semiglobal: methods return true and LogProb::ln_one().

Required Methods

fn free_start_gap_x(&self) -> bool

Allow free start gap in x.

fn free_end_gap_x(&self) -> bool

Allow free end gap in x.

Provided Methods

fn prob_start_gap_x(&self, i: usize) -> LogProb

Probability to start at x[i]. This can be left unchanged if you use free_start_gap_x and free_end_gap_x.