Struct bio::stats::probs::cdf::CDF

pub struct CDF<T: Ord> { /* fields omitted */ }

Implementation of a cumulative distribution function.

Methods

impl<T: Ord> CDF<T>

pub fn from_pmf(entries: Vec<Entry<T>>) -> Self

Create CDF from given probability mass function (PMF). The PMF may contain duplicate values the probabilities of which are summed during generation of the CDF.

Arguments

pmf - the PMF as a vector of Entry objects

pub fn from_cdf<I: Iterator<Item = Entry<T>>>(entries: I) -> Self

Create CDF from iterator. This can be used to replace the values of a CDF.

pub fn reduce(self) -> Self

Reduce CDF by omitting values with zero probability.

pub fn sample(self, n: usize) -> Self

Downsample CDF to n entries. Panics if n <= 1 and returns identity if n is greater than the number of entries.

pub fn iter(&self) -> Iter<Entry<T>>

Provide iterator.

pub fn iter_mut(&mut self) -> IterMut<Entry<T>>

Mutable iterator over entries. This does not check for consistency. In other words, you should not change the order of the entries, nor the probabilities!

pub fn iter_pmf(&self) -> CDFPMFIter<T>

Iterator over corresponding PMF.

pub fn get(&self, value: &T) -> Option<LogProb>

Get cumulative probability for a given value. If the value is not present, return the probability of the previous value. Complexity O(log n).

pub fn get_pmf(&self, value: &T) -> Option<LogProb>

Get probability (i.e. probability mass) for a given value. Complexity O(log n).

pub fn total_prob(&self) -> LogProb

Return total probability.

pub fn map(&self) -> Option<&T>

Return maximum a posteriori probability estimate (MAP).

pub fn credible_interval(&self, width: f64) -> Option<Range<&T>>

Return w%-credible interval. The width w is a float between 0 and 1. Panics otherwise. E.g. provide width=0.95 for the 95% credible interval.

pub fn len(&self) -> usize

Number of entries in the CDF.

pub fn is_empty(&self) -> bool

impl<T: Clone + Ord> CDF<T> where
f64: From<T>,

pub fn expected_value(&self) -> f64

Calculate expected value.

pub fn variance(&self) -> f64

Calculate variance.

pub fn standard_deviation(&self) -> f64

Calculate standard deviation.