Struct bio::stats::probs::LogProb

pub struct LogProb(pub f64);

A newtype for log-scale probabilities. For performance reasons, we use an approximation of the exp method implemented in bio::utils::FastExp. This can lead to slight errors, but should not matter given that most of the probability differences are reflected within the integer part of the log probability.

Example

#[macro_use]
extern crate approx;
use bio::stats::{LogProb, Prob};

// convert from probability
let p = LogProb::from(Prob(0.5));
// convert manually
let q = LogProb(0.2f64.ln());
// obtain zero probability in log-space
let o = LogProb::ln_one();

assert_relative_eq!(*Prob::from(p.ln_add_exp(q) + o), *Prob(0.7), epsilon=0.000001);

Implementations

impl LogProb

pub fn is_valid(&self) -> bool

pub fn ln_zero() -> LogProb

Log-space representation of Pr=0

pub fn ln_one() -> LogProb

Log-space representation of Pr=1

pub fn cap_numerical_overshoot(&self, epsilon: f64) -> LogProb

sums of LogProbs, e.g. with ln_sum_exp() can end up slightly above the maximum of LogProb <= 0 due to numerical imprecisions -- this function can rescue such values before panics due to asserts in other functions handling LogProbs, e.g. ln_1m_exp

pub fn ln_one_minus_exp(&self) -> LogProb

Numerically stable calculation of 1 - p in log-space.

pub fn ln_sum_exp(probs: &[LogProb]) -> LogProb

Numerically stable sum of probabilities in log-space.

pub fn ln_add_exp(self, other: LogProb) -> LogProb

Numerically stable addition of probabilities in log-space.

pub fn ln_sub_exp(self, other: LogProb) -> LogProb

Numerically stable subtraction of probabilities in log-space.

pub fn ln_cumsum_exp<I: IntoIterator<Item = LogProb>>(probs: I) -> ScanIter<I>

Calculate the cumulative sum of the given probabilities in a numerically stable way (Durbin 1998).

pub fn ln_trapezoidal_integrate_exp<T, D>(
density: D,
a: T,
b: T,
n: usize
) -> LogProb where
T: Copy + Add<Output = T> + Sub<Output = T> + Div<Output = T> + Mul<Output = T> + Float,
D: FnMut(usize, T) -> LogProb,
f64: From<T>,

Integrate numerically stable over given log-space density in the interval [a, b]. Uses the trapezoidal rule with n grid points.

pub fn ln_simpsons_integrate_exp<T, D>(
density: D,
a: T,
b: T,
n: usize
) -> LogProb where
T: Copy + Add<Output = T> + Sub<Output = T> + Div<Output = T> + Mul<Output = T> + Float,
D: FnMut(usize, T) -> LogProb,
f64: From<T>,

Integrate numerically stable over given log-space density in the interval [a, b]. Uses Simpson's rule with n (odd) grid points.