Module bio::utils

Common utilities.

Structs

Interval
An Interval wraps the std::ops::Range from the stdlib and is defined by a start and end field where end should be >= start.

Traits

FastExp
This trait adds a fast approximation of exp to float types.

Functions

prescan

scan
In place implementation of scan over a slice.

trim_newline
Remove a trailing newline from the given string in place.

Type Definitions

Text
Type alias for an owned text, i.e. Vec<u8>.

TextSlice
Type alias for a text slice, i.e. &[u8].