Struct bio::utils::Interval

pub struct Interval<N: Ord + Clone>(_);

An Interval wraps the std::ops::Range from the stdlib and is defined by a start and end field where end should be >= start.

Methods

impl<N: Ord + Clone> Interval<N>

fn new(r: Range<N>) -> Result<Interval<N>, IntervalError>

Construct a new Interval from the given Range. Will return Err if end < start.