Crate bio

Rust-bio, a rusty bioinformatics library.

This library provides implementations of many algorithms and data structures that are useful for bioinformatics. All provided implementations are rigorously tested via continuous integration (https://travis-ci.org/johanneskoester/rust-bio).

Currently, rust-bio provides

pattern matching,

pairwise alignment,

suffix arrays,

BWT and FM-Index,

rank/select data structures.

Find out more at https://github.com/johanneskoester/rust-bio.

Modules

alignment
Alignment algorithms.

alphabets
Handling different alphabets.

bitencoding
A fixed-width bit encoding implementation.

data_structures
Various useful data structures.

io
Parsers for important file formats.

pattern_matching
This module contains various useful pattern matching algorithms. The implementations are based on the lecture notes "Algorithmen auf Sequenzen", Kopczynski, Marschall, Martin and Rahmann, 2008 - 2015.

utils
Common utilities.