Module bio::alignment

Various alignment and distance computing algorithms.

Modules

distance
Various subroutines for computing a distance between sequences. Complexity: O(n) for strings of length n for the Hamming distance; O(n * m) for strings of length n and m for the Levenshtein (or edit) distance.

pairwise
Calculate alignments with a generalized variant of the Smith Waterman algorithm. Complexity: O(n * m) for strings of length m and n.

Structs

Alignment
An alignment, consisting of a score, the start and end position of the alignment on sequence x and sequence y, the length of sequence x, and the alignment edit operations (see alignment::pairwise for meaning of x and y).

Enums

AlignmentOperation
Alignment operations (Match, Subst, Del and Ins).