Module bio::alignment::distance::simd

String distance routines accelerated with Single Instruction Multiple Data (SIMD) intrinsics.

These routines will automatically fallback to scalar versions if AVX2 or SSE4.1 is not supported by the CPU.

With AVX2, SIMD-accelerated Hamming distance can reach up to 40 times faster than the scalar version on strings that are long enough.

The performance of SIMD-accelerated Levenshtein distance depends on the number of edits between two strings, so it can perform anywhere from 2 times to nearly 1000 times faster than the scalar version. When the two strings are completely different, there could be no speedup at all. It is important to note that the algorithms work best when the number of edits is known to be small compared to the length of the strings (for example, 10% difference). This should be applicable in many situations.

If AVX2 support is not available, there is a speed penalty for using SSE4.1 with smaller vectors.

Functions

bounded_levenshtein
SIMD-accelerated bounded Levenshtein (or Edit) distance between two strings. Complexity: O(k / w * (n + m)), with n and m being the length of the given texts, k being the threshold on the number of edits, and w being the length of the SIMD vectors (usually w = 16 or w = 32).

hamming
SIMD-accelerated Hamming distance between two strings. Complexity: O(n / w), for SIMD vectors of length w (usually w = 16 or w = 32).

levenshtein
SIMD-accelerated Levenshtein (or Edit) distance between two strings. Complexity: O(k / w * (n + m)), with n and m being the length of the given texts, k being the number of edits, and w being the length of the SIMD vectors (usually w = 16 or w = 32).