Module bio::alignment::pairwise

Calculate alignments with a generalized variant of the Smith Waterman algorithm. Complexity: O(n * m) for strings of length m and n.

Example

use bio::alignment::pairwise::Aligner;
use bio::alignment::AlignmentOperation::{Match, Subst, Ins, Del};
let x = b"ACCGTGGAT";
let y = b"AAAAACCGTTGAT";
let score = |&: a: u8, b: u8| if a == b {1i32} else {-1i32};
let mut aligner = Aligner::with_capacity(x.len(), y.len(), -5, -1, score);
let alignment = aligner.semiglobal(x, y);
assert_eq!(alignment.i, 4);
assert_eq!(alignment.j, 0);
assert_eq!(alignment.operations, [Match, Match, Match, Match, Match, Subst, Match, Match, Match]);

Structs

Aligner