Struct bio::alignment::Alignment

pub struct Alignment {
    pub score: i32,
    pub ystart: usize,
    pub xstart: usize,
    pub xlen: usize,
    pub operations: Vec<AlignmentOperation>,
}

An alignment, consisting of a score, a start in sequence y, a start in sequence x, a length and its edit operations (see alignment::pairwise for meaning of x and y).

Fields

score

ystart

xstart

xlen

operations

Methods

impl Alignment

fn cigar(&self, hard_clip: bool) -> String

Calculate the cigar string.