Module bio::data_structures::bwt

The Burrows-Wheeler-Transform and related data structures. The implementation is based on the lecture notes “Algorithmen auf Sequenzen”, Kopczynski, Marschall, Martin and Rahmann, 2008 - 2015.

Structs

Occ

An occurrence array implementation.

Functions

bwt

Calculate Burrows-Wheeler-Transform of the given text of length n. Complexity: O(n).

bwtfind

Calculate the bwtfind array needed for inverting the BWT. Complexity O(n).

invert_bwt

Calculate the inverse of a BWT of length n, which is the original text. Complexity: O(n).

less

Calculate the less array for a given BWT. Complexity O(n).

Type Definitions

BWT

BWTFind

BWTSlice

Less