Module bio::data_structures::fmindex

FM-Index and FMD-Index for finding suffix array intervals matching a given pattern in linear time.

Structs

BiInterval
A bi-interval on suffix array of the forward and reverse strand of a DNA text.

FMDIndex
The FMD-Index for linear time search of supermaximal exact matches on forward and reverse strand of DNA texts (Li, 2012).

FMIndex
The Fast Index in Minute space (FM-Index, Ferragina and Manzini, 2000) for finding suffix array intervals matching a given pattern.

Interval
A suffix array interval.

Traits

FMIndexable