Struct bio::data_structures::fmindex::FMDIndex

pub struct FMDIndex<'a> {
    // some fields omitted
}

The FMD-Index for linear time search of supermaximal exact matches on forward and reverse strand of DNA texts (Li, 2012).

Methods

impl<'a> FMDIndex<'a>

fn new(bwt: &'a BWT, k: usize) -> Self

Construct a new instance of the FMD index (see Heng Li (2012) Bioinformatics). This expects a BWT that was created from a text over the DNA alphabet with N (alphabets::dna::n_alphabet()) consisting of the concatenation with its reverse complement, separated by the sentinel symbol $. I.e., let T be the original text and R be its reverse complement. Then, the expected text is T$R$. Further, multiple concatenated texts are allowed, e.g. T1$R1$T2$R2$T3$R3$.

Arguments

bwt - the BWT

k - the sampling rate of the occ array: every k-th entry will be stored (higher k means less memory usage, but worse performance)

fn smems(&self, pattern: &[u8], i: usize) -> Vec<BiInterval>

Find supermaximal exact matches of given pattern that overlap position i in the pattern. Complexity O(m) with pattern of length m.

Example

use bio::data_structures::fmindex::FMDIndex;
use bio::data_structures::suffix_array::suffix_array;
use bio::data_structures::bwt::bwt;

let text = b"ATTC$GAAT$";
let pos = suffix_array(text);
let bwt = bwt(text, &pos);
let fmdindex = FMDIndex::new(&bwt, 3);

let pattern = b"ATT";
let intervals = fmdindex.smems(pattern, 2);
let occ = intervals[0].occ(&pos);
let occ_revcomp = intervals[0].occ_revcomp(&pos);

assert_eq!(occ, [0]);
assert_eq!(occ_revcomp, [6]);