Module bio::data_structures::interval_tree

Interval tree, a data structure for efficiently storing and searching numeric intervals.

This data structure uses the std::ops:Range type to define the intervals. The range bounds may be specified by any type from the num crate that satisfies the std::cmp::Ord trait. Upon inserting an interval may be associated with a data value. The interval data is stored in an augmented AVL-tree which allows for efficient inserting and querying.

Example

use bio::data_structures::interval_tree::IntervalTree;

let mut tree = IntervalTree::new();
tree.insert((11..20), "Range_1").unwrap();
tree.insert((25..30), "Range_2").unwrap();
for r in tree.find(&(15..25)).unwrap() {
    assert_eq!(r.interval(), &(11..20));
    assert_eq!(r.data(), &"Range_1");
}Run

Structs

Entry
An Entry is used by the IntervalTreeIterator to return references to the data in the tree

IntervalTree
An interval tree for storing Ranges with data

IntervalTreeIterator
An IntervalTreeIterator is returned from by find and iterates over the entries overlapping the query

Enums

IntervalTreeError