Struct bio::data_structures::qgram_index::QGramIndex

pub struct QGramIndex {
    // some fields omitted
}

A classical, flexible, q-gram index implementation.

Methods

impl QGramIndex

fn new(q: u32, text: &[u8], alphabet: &Alphabet) -> QGramIndex

Create a new q-gram index. The q has to be smaller than b / log2(|A|) with |A| being the alphabet size and b the number bits with the usize data type.

fn with_max_count(q: u32, text: &[u8], alphabet: &Alphabet, max_count: usize) -> QGramIndex

Create a new q-gram index, only considering q-grams that occur at most max_count times. The q has to be smaller than b / log2(|A|) with |A| being the alphabet size and b the number bits with the usize data type.

fn qgram_matches(&self, qgram: usize) -> &[usize]

Return text positions with matching q-gram.

fn matches(&self, pattern: &[u8], min_count: usize) -> Vec<Match>

Return matches of the given pattern.

fn exact_matches(&self, pattern: &[u8]) -> Vec<ExactMatch>

Return exact matches of the given pattern.