Module bio::data_structures::suffix_array

Suffix arrays and related algorithms.

Functions

get_lcp
Construct lcp array for given text and suffix array of length n. Complexity: O(n).

get_suffix_array
Construct suffix array for given text of length n. Complexity: O(n). This is an implementation of the induced sorting as presented by Ge Nong, Sen Zhang und Wai Hong Chan (2009), also known as SAIS. The implementation is based on the following lecture notes: http://ls11-www.cs.tu-dortmund.de/people/rahmann/algoseq.pdf

Type Definitions

LCPArray

SuffixArray