Struct bio::io::bed::Record

pub struct Record {
    // some fields omitted
}

A BED record as defined by BEDtools (http://bedtools.readthedocs.org/en/latest/content/general-usage.html)

Methods

impl Record

fn new() -> Self

Create a new BED record.

fn chrom(&self) -> &str

Chromosome of the feature.

fn start(&self) -> u64

Start position of feature (0-based).

fn end(&self) -> u64

End position of feature (0-based, not included).

fn name(&self) -> Option<&str>

Name of the feature.

fn score(&self) -> Option<&str>

Score of the feature.

fn strand(&self) -> Option<Strand>

Strand of the feature.

fn aux(&self, i: usize) -> Option<&str>

Access auxilliary fields after the strand field by index (counting first field (chromosome) as 0).

fn set_chrom(&mut self, chrom: &str)

Set chromosome.

fn set_start(&mut self, start: u64)

Set start of feature.

fn set_end(&mut self, end: u64)

Set end of feature.

fn set_name(&mut self, name: &str)

Set name.

fn set_score(&mut self, score: &str)

Set score.

fn push_aux(&mut self, field: &str)

Add auxilliary field. This has to happen after name and score have been set.