Module bio::io::fasta

Structs and trait to read and write files in FASTA format.

Example

Read

In this example, we parse a fasta file from stdin and compute some statistics

use std::io;
use bio::io::fasta;

let mut reader = fasta::Reader::new(io::stdin());

let mut nb_reads = 0;
let mut nb_bases = 0;

for result in reader.records() {
    let record = result.expect("Error during fasta record parsing");
    println!("{}", record.id());

    nb_reads += 1;
    nb_bases += record.seq().len();
}

println!("Number of reads: {}", nb_reads);
println!("Number of bases: {}", nb_bases);

We can also use a while loop to iterate over records. This is slightly faster than the for loop.

use std::io;
use bio::io::fasta;
let mut records = fasta::Reader::new(io::stdin()).records();

let mut nb_reads = 0;
let mut nb_bases = 0;

while let Some(Ok(record)) = records.next() {
    nb_reads += 1;
    nb_bases += record.seq().len();
}

println!("Number of reads: {}", nb_reads);
println!("Number of bases: {}", nb_bases);

Write

In this example we generate 10 random sequences with length 100 and write them to stdout.

use std::io;
use bio::io::fasta;

let mut seed = 42;

let nucleotides = [b'A', b'C', b'G', b'T'];

let mut writer = fasta::Writer::new(io::stdout());

for _ in 0..10 {
    let seq = (0..100).map(|_| {
        seed = ((seed ^ seed << 13) ^ seed >> 7) ^ seed << 17; // don't use this random generator
        nucleotides[seed % 4]
    }).collect::<Vec<u8>>();

   writer.write("random", None, seq.as_slice()).ok().expect("Error writing record.");
}

Read and Write

In this example we filter reads from stdin on sequence length and write them to stdout

use std::io;
use bio::io::fasta;
use bio::io::fasta::FastaRead;

let mut reader = fasta::Reader::new(io::stdin());
let mut writer = fasta::Writer::new(io::stdout());
let mut record = fasta::Record::new();

while let Ok(()) = reader.read(&mut record) {
    if record.is_empty() {
        break;
    }

    if record.seq().len() > 100 {
        writer.write_record(&record).ok().expect("Error writing record.");
    }
}

Structs

Index
A FASTA index as created by SAMtools (.fai).

IndexedReader
A FASTA reader with an index as created by SAMtools (.fai).

IndexedReaderIterator

Reader
A FASTA reader.

Record
A FASTA record.

Records
An iterator over the records of a Fasta file.

Sequence
A sequence record returned by the FASTA index.

Writer
A Fasta writer.

Traits

FastaRead
Trait for FASTA readers.