Module bio::pattern_matching

This module contains various useful pattern matching algorithms. The implementations are based on the lecture notes "Algorithmen auf Sequenzen", Kopczynski, Marschall, Martin and Rahmann, 2008 - 2015.

Algorithm of Horspool: fastest for a sufficiently large alphabet

Shift And algorithm: fast for patterns with less than 64 symbols and very small alphabets.

BNDM algorithm: fast for patterns with less than 64 symbols.

BOM algorithm: fast for long patterns and small alphabet.

KMP algorithm: the classic.

Modules

bndm
Backward nondeterministic DAWG matching (BNDM). Best-case complexity: O(n / m) with pattern of length m <= 64 and text of length n. Worst case complexity: O(n * m).

bom
Backward oracle matching algorithm.

Example

horspool
Algorithm of Horspool. Window-based, similar to but faster than Boyer-Moore.

kmp
Algorithm of Knuth Morris and Pratt. Constructs an automaton recognizing the pattern, and scans linearly over a text of length n. Complexity: O(n). Here, the automaton is implemented by directly storing the transition function delta in a table for each state and symbol in the alphabet. For this, it uses the lps-function, that assigns to each position q in the pattern the longest prefix of the pattern that is suffix of pattern[..q+1]. Then, in the NFA for the pattern, active states after reading position q are {q, lps(q), lps(lps(q)), ... 0}.

shift_and
ShiftAnd algorithm for pattern matching. Patterns may contain at most 64 symbols. Complexity: O(n) with text length n.