Module bio::pattern_matching::myers::long

Block-based version of the algorithm, which does not restrict pattern length.

This module implements the block-based version of the Myers pattern matching algorithm. It can be used for searching patterns of any length and obtaining semiglobal alignments of the hits. Apart from that, the Myers object in this module provides exactly the same API as the ‘simple’ version bio::pattern_matching::myers::Myers. For short patterns, the ‘simple’ version is still to be preferred, as the block-based algorithm is slower.

Structs

FullMatches

Iterator over tuples of starting position, end position and distance of matches. In addition, methods for obtaining the hit alignment path are provided.

LazyMatches

Iterator over tuples of end position and distance of matches. In addition, methods for obtaining the hit alignment path are provided.

Matches

Iterator over pairs of end positions and distance of matches.

Myers

Myers algorithm.