Trait bio::pattern_matching::myers::BitVec

pub trait BitVec: Copy + Debug + Binary + Default + Add + Sub + BitOr + BitOrAssign + BitAnd + BitXor + Not + Shl<usize> + ShlAssign<usize> + ShrAssign<usize> + PrimInt + WrappingAdd + ToPrimitive + FromPrimitive + AsPrimitive<usize> + AsPrimitive<i64> {
    type DistType: DistType;
}

This trait must be implemented for integer types serving as bit vectors. Only unsigned integers will work correctly.

Associated Types

type DistType: DistType

Type that should be used to store the distance score when using the simple Myers algorithm (not the block-based one, which always uses usize).

Loading content...