Module bio::stats

Mathematical and statistical tools.

Re-exports

pub use crate::stats::probs::LogProb;

pub use crate::stats::probs::PHREDProb;

pub use crate::stats::probs::Prob;

Modules

bayesian
Utilities for Bayesian statistics.

combinatorics
Combinations with and without replacement.

hmm
An implementation of Hidden Markov Models in Rust.

pairhmm
A pair Hidden Markov Model for calculating the probability that two sequences are related to each other. Depending on the used parameters, this can, e.g., be used to calculate the probability that a certain sequencing read comes from a given position in a reference genome.

probs
Handling log-probabilities. Log probabilities are an important tool to deal with probabilities in a numerically stable way, in particular when having probabilities close to zero.