Module bio::stats::bayesian::model

A trait system for Bayesian statistical modelling.

Structs

Model
Bayesian model, consisting of a prior, a posterior and a likelihood model. Thereby, Payload is a custom payload of the model instance. This can be used to define custom caching mechanisms. See here for an example.

ModelInstance
Instance of a model for given data and event universe. From the instance, posterior, marginal and MAP can be computed.

Traits

Likelihood
Likelihood model.

Posterior
Posterior model.

Prior
Prior model.

Type Definitions

JointProbUniverse