Module bio::stats::probs

Handling log-probabilities. Log probabilities are an important tool to deal with probabilities in a numerically stable way, in particular when having probabilities close to zero.

Re-exports

pub use self::errors::Error;

pub use self::errors::Result;

Modules

adaptive_integration

cdf

Support for discrete probability distributions in terms of cumulative distribution functions (CDF).

errors

Error definitions for the probs module.

Structs

LogProb

A newtype for log-scale probabilities. For performance reasons, we use an approximation of the exp method implemented in bio::utils::FastExp. This can lead to slight errors, but should not matter given that most of the probability differences are reflected within the integer part of the log probability.

PHREDProb

A newtype for PHRED-scale probabilities.

Prob

A newtype for probabilities.

Type Definitions

ScanIter

Iterator returned by scans over logprobs.