Function bio::stats::probs::adaptive_integration::ln_integrate_exp

pub fn ln_integrate_exp<T, F>(
    density: F, 
    min_point: T, 
    max_point: T, 
    max_resolution: T
) -> LogProb where
    T: Copy + Add<Output = T> + Sub<Output = T> + Div<Output = T> + Div<NotNan<f64>, Output = T> + Mul<Output = T> + Into<f64> + From<f64> + Ord + Debug + Hash,
    F: FnMut(T) -> LogProb,
    f64: From<T>, 

Integrate over an interval of type T with a given density function while trying to minimize the number of grid points evaluated and still hit the maximum likelihood point. This is achieved via a binary search over the grid points. The assumption is that the density is unimodal. If that is not the case, the binary search will not find the maximum and the integral can miss features.

Example

use bio::stats::probs::adaptive_integration::ln_integrate_exp;
use bio::stats::probs::{Prob, LogProb};
use statrs::distribution::{Normal, Continuous};
use statrs::statistics::Distribution;
use ordered_float::NotNan;
use approx::abs_diff_eq;

let ndist = Normal::new(0.0, 1.0).unwrap();

let integral = ln_integrate_exp(
    |x| LogProb::from(Prob(ndist.pdf(*x))),
    NotNan::new(-1.0).unwrap(),
    NotNan::new(1.0).unwrap(),
    NotNan::new(0.01).unwrap()
);
abs_diff_eq!(integral.exp(), 0.682, epsilon=0.01);