Module bio::utils

Common utilities.

Structs

NonNaNFloat

Traits

IntoTextIterator
Type alias for a type that can be coerced into a TextIterator. This includes &Vec<u8>, &[u8], Iterator<Item=&u8>.

TextIterator
Type alias for an iterator over a sequence, i.e. Iterator<Item=&u8>.

Functions

prescan

scan
In place implementation of scan over a slice.

trim_newline
Remove a trailing newline from the given string in place.

Type Definitions

Text
Type alias for an owned text, i.e. Vec<u8>.

TextSlice
Type alias for a text slice, i.e. &[u8].